# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Glory Sports API - Cricket sports ecosystem for India
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{userId}/ban:
    patch:
      operationId: adminBanUser
      tags: [admin]
      summary: Admin - ban or unban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserInput"
      responses:
        "200":
          description: User ban status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/turfs:
    get:
      operationId: adminListTurfs
      tags: [admin]
      summary: Admin - list all turfs pending verification
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Turfs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Turf"
  /admin/turfs/{turfId}/verify:
    patch:
      operationId: adminVerifyTurf
      tags: [admin]
      summary: Admin - verify or reject a turf
      parameters:
        - name: turfId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTurfInput"
      responses:
        "200":
          description: Turf verification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turf"
  /admin/matches:
    get:
      operationId: adminListMatches
      tags: [admin]
      summary: Admin - list matches with optional suspicious filter
      parameters:
        - name: suspicious
          in: query
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
  /admin/matches/{matchId}/flag:
    patch:
      operationId: adminFlagMatch
      tags: [admin]
      summary: Admin - flag or clear a suspicious match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagMatchInput"
      responses:
        "200":
          description: Match flagged status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, phone, password, role]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        password:
          type: string
        role:
          type: string
          enum: [player, turf_owner]
        city:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [player, turf_owner, admin]
        city:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isBanned:
          type: boolean
        isVerified:
          type: boolean
        createdAt:
          type: string

    # PLAYERS
...
    Turf:
      type: object
      required: [id, ownerId, name, city, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
        name:
          type: string
        city:
          type: string
        address:
          type: ["string", "null"]
        sports:
          type: array
          items:
            type: string
        pricePerHour:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        photos:
          type: array
          items:
            type: string
        verificationStatus:
          type: string
          enum: [pending, verified, rejected]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Match:
      type: object
      required: [id, turfId, teamAId, teamBId, status, createdAt]
      properties:
        id:
          type: integer
        turfId:
          type: integer
        turfName:
          type: string
        teamAId:
          type: integer
        teamAName:
          type: string
        teamBId:
          type: integer
        teamBName:
          type: string
        overs:
          type: integer
        matchType:
          type: string
        matchDate:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, verified, disputed, flagged]
        teamAScore:
          type: ["integer", "null"]
        teamAWickets:
          type: ["integer", "null"]
        teamBScore:
          type: ["integer", "null"]
        teamBWickets:
          type: ["integer", "null"]
        winnerTeamId:
          type: ["integer", "null"]
        mvpPlayerId:
          type: ["integer", "null"]
        mvpPlayerName:
          type: ["string", "null"]
        teamACaptainVerified:
          type: boolean
        teamBCaptainVerified:
          type: boolean
        isSuspicious:
          type: boolean
        bookingId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    AdminUserListResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
    BanUserInput:
      type: object
      required: [isBanned]
      properties:
        isBanned:
          type: boolean
        reason:
          type: string
    VerifyTurfInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [verified, rejected]
        notes:
          type: string
    FlagMatchInput:
      type: object
      required: [isSuspicious]
      properties:
        isSuspicious:
          type: boolean
        reason:
          type: string