# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Glory Sports API - Cricket sports ecosystem for India
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  role: UserRole;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isBanned?: boolean;
  isVerified?: boolean;
  createdAt: string;
}